backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "black>=25.9.0",
    "email-validator>=2.0.0",
    "fastapi>=0.121.0",
    "isort>=7.0.0",
    "bcrypt>=4.0.1",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "ruff>=0.14.4",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
